Add files via upload
Considering three numbers provided by the user as length of sides of a triangle, first check if the values are valid for representing the sides of a triangle (i.e. whether a triangle can be drawn using the given values). If the lengths of sides are valid, print the type of the triangle.